Removed instrumentation.

+++ NTDLS.Katzebase.Engine/Atomicity/TransactionSnapshot.cs
@@ -32,6 +32,6 @@ namespace NTDLS.Katzebase.Engine.Atomicity
         public DeferredDiskIOSnapshot DeferredIOs { get; set; } = new();
         public bool IsCommittedOrRolledBack { get; set; } = false;
         public bool IsCancelled { get; set; } = false;
-        public int ReferenceCount { get; set; }
+        public long ReferenceCount { get; set; }
     }
 }
